[CORDA-1937]: Fixing internal API exposures.
