+++ node-api/src/main/kotlin/net/corda/nodeapi/config/CertificateStoreSupplier.kt
@@ -1,0 +1,18 @@
+package net.corda.nodeapi.config
+
+import java.io.IOException
+
+// TODO sollecitom see if you can make this private API wise
+interface CertificateStoreSupplier {
+
+    fun get(createNew: Boolean = false): CertificateStore
+
+    fun getOptional(): CertificateStore? {
+
+        return try {
+            get()
+        } catch (e: IOException) {
+            null
+        }
+    }
+}
