+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/config/SSLConfiguration.kt
@@ -1,60 +1,16 @@
 package net.corda.nodeapi.internal.config
 
 import net.corda.core.internal.div
-import net.corda.core.internal.outputStream
-import net.corda.nodeapi.internal.crypto.CertificateAndKeyPair
+import net.corda.nodeapi.config.CertificateStoreSupplier
+import net.corda.nodeapi.config.CertificateStore
 import net.corda.nodeapi.internal.crypto.X509KeyStore
-import java.io.IOException
-import java.io.OutputStream
-import java.nio.file.OpenOption
 import java.nio.file.Path
-import java.security.cert.X509Certificate
 
+// TODO sollecitom use this
 interface SslConfiguration {
 
     val keyStore: CertificateStoreSupplier
     val trustStore: CertificateStoreSupplier
-}
-
-// TODO sollecitom see if you can make the password private here
-interface CertificateStore {
-
-    // TODO sollecitom ideally make this private
-    val value: X509KeyStore
-    // TODO sollecitom see if this can stay private (by adding delegate functions over X509Store)
-    val password: String
-
-    fun writeTo(stream: OutputStream) = value.internal.store(stream, password.toCharArray())
-
-    fun writeTo(path: Path, vararg options: OpenOption) = path.outputStream(*options)
-
-    fun <RESULT> update(action: X509KeyStore.() -> RESULT): RESULT {
-        val result = action.invoke(value)
-        value.save()
-        return result
-    }
-
-    fun getCertificate(alias: String): X509Certificate = value.getCertificate(alias)
-
-    operator fun contains(alias: String): Boolean = value.contains(alias)
-
-    fun getCertificateChain(alias: String): List<X509Certificate> = value.getCertificateChain(alias)
-
-    fun getCertificateAndKeyPair(alias: String, keyPassword: String = password): CertificateAndKeyPair = value.getCertificateAndKeyPair(alias, keyPassword)
-}
-
-interface CertificateStoreSupplier {
-
-    fun get(createNew: Boolean = false): CertificateStore
-
-    fun getOptional(): CertificateStore? {
-
-        return try {
-            get()
-        } catch (e: IOException) {
-            null
-        }
-    }
 }
 
 class FileBasedCertificateStoreLoader(private val path: Path, private val password: String) : CertificateStoreSupplier {
